+++ app/libraries/Jentil/Utilities/ThemeMods/Footer.php
@@ -72,12 +72,12 @@ class Footer extends AbstractThemeMod
 
         return \sprintf(
             \esc_html__('&copy; %1$s %2$s. Built with %3$s on %4$s', 'jentil'),
-            '<span itemprop="copyrightYear">{{this_year}}</span>',
-            '<a class="blog-name" itemprop="url" href="{{site_url}}">'.
-                '<span itemprop="copyrightHolder">{{site_name}}</span></a>',
-            '<em><a itemprop="url" rel="nofollow" href="'.
-                $meta['theme_uri'].'">'.$meta['name'].'</a></em>',
-            '<a itemprop="url" rel="nofollow" href="https://wordpress.org">WordPress</a>.'
+            '{{this_year}}',
+            '<a class="blog-name" href="{{site_url}}">{{site_name}}</a>',
+            '<em><a rel="nofollow" href="'.$meta['uri'].'">'.
+                $meta['name'].
+            '</a></em>',
+            '<a rel="nofollow" href="https://wordpress.org">WordPress</a>.'
         );
     }
 }