+++ app/partials/header-mini.php
@@ -3,7 +3,7 @@ declare (strict_types = 1);
 
 \Jentil()->utilities->loader->loadPartial('header', 'micro'); ?>
 
-<header id="header" class="site-header" itemscope itemtype="http://schema.org/WPHeader">
+<header id="header" class="site-header">
     <?php \do_action('jentil_inside_header'); ?>
 </header><!-- #header -->