+++ app/templates/singular.php
@@ -10,17 +10,14 @@ declare (strict_types = 1);
         \the_ID();
     ?>" id="post-<?php \the_ID(); ?>" <?php
         \post_class(['post-wrap']);
-    ?> itemscope itemtype="http://schema.org/Article">
+    ?>>
         <?php if ($post->post_title) { ?>
             <header class="page-header">
         <?php }
 
         \do_action('jentil_before_title');
 
-        \the_title(
-            '<h1 class="entry-title" itemprop="name headline mainEntityOfPage">',
-            '</h1>'
-        );
+        \the_title('<h1 class="entry-title">', '</h1>');
 
         \do_action('jentil_after_title');
 
@@ -30,7 +27,7 @@ declare (strict_types = 1);
 
         \do_action('jentil_before_content'); ?>
 
-        <div class="entry-content" itemprop="articleBody">
+        <div class="entry-content">
             <?php \the_content();
 
             \wp_link_pages([
@@ -38,7 +35,7 @@ declare (strict_types = 1);
                 . \esc_html__('Pages: ', 'jentil'),
                 'after' => '</nav>',
             ]); ?>
-        </div><!-- .entry-content -->
+        </div>
 
         <?php \do_action('jentil_after_content'); ?>
     </article>