+++ assets/scripts/core.ts
@@ -1,30 +1,9 @@
-/*!
- * Jentil
- *
- * @author [GrottoPress](https://www.grottopress.com)
- * @author [N Atta Kusi Adusei](https://twitter.com/akadusei)
- */
+/// <reference path='./core/module.d.ts' />
 
-/// <reference path='./global.d.ts' />
+import { Base } from './core/base'
 
-namespace Jentil
-{
-    export class App
-    {
-        public constructor(private readonly _j: JQueryStatic)
-        {
-        }
+import { BodyClass } from './core/body-class'
 
-        public run(): void
-        {
-            this.addBodyClasses()
-        }
+const cores = [new BodyClass(jQuery)]
 
-        private addBodyClasses(): void
-        {
-            this._j('body').removeClass('no-js').addClass('has-js')
-        }
-    }
-}
-
-new Jentil.App(jQuery).run()
+jQuery.each(cores, (_, core: Base): void => core.run())