+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.11.0] - 2020-02-08
+
+### Added
+- Add support for PHP 7.4
+- Set up JS module bundling with [rollup js](https://rollupjs.org)
+- Add `Procfile`
+
+### Changed
+- Remove `theme_` prefix from theme meta keys
+
+### Removed
+- Remove HTML microdata
+
 ## 0.10.2 - 2019-04-30
 
 ### Added